Everything implemented.
Still a WiP, however all the memory store functionality has now been
implemented in the mongodb store. Just need to support configuration of the
server and test, test, test.

I've changed the semantics slightly, if a session id is given that hasn't been
seen before a new session is created with a *NEW* session id, rather than a new
one with the given id, hopefully this won't confuse things.... i had to do this
as my session ids are intrinsic to the db so I Can't control the construction of them
I also don't like the idea of having an external client being able to set the id of
things on the server, it feels wrong, but its probably nothing to sleep over ;)

The mongoDB driver is also of concern, there are various places where I can see
we're at risk of our callbacks being 'called-back-to' repeatedly, where they should
only be called back to once, not sure what to do about this, fix the driver, or guard
the code ! :(